+++ utils/utils.go
@@ -46,6 +46,12 @@ const (
 	ErrorStatusSharedCredentialUnverified string = "shared-credential-unverified"
 	//ErrorStatusNotAllowed ...
 	ErrorStatusNotAllowed string = "not-allowed"
+
+	//Character sets for password generation
+	upper   string = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+	lower   string = "abcdefghijklmnopqrstuvwxyz"
+	digits  string = "0123456789"
+	special string = "!@#$%^&*()"
 )
 
 // SetRandomSeed sets the seed for random number generation
@@ -80,6 +86,19 @@ func GenerateRandomString(s int) (string, error) {
 // GenerateRandomInt returns a random integer between 0 and max
 func GenerateRandomInt(max int) int {
 	return rand.Intn(max)
+}
+
+// GenerateRandomPassword returns a randomly generated password string
+func GenerateRandomPassword(s int) string {
+	validCharacters := []byte(upper + lower + digits + special)
+	rand.Shuffle(len(validCharacters), func(i, j int) { validCharacters[i], validCharacters[j] = validCharacters[j], validCharacters[i] })
+
+	password := make([]byte, s)
+	for i := 0; i < s; i++ {
+		password[i] = validCharacters[rand.Intn(len(validCharacters))]
+	}
+
+	return string(password)
 }
 
 // ConvertToJSON converts to json
@@ -152,6 +171,26 @@ func Contains(list []string, value string) bool {
 		}
 	}
 	return false
+}
+
+//StringListDiff returns a list of added, removed, unchanged values between a new and old string list
+func StringListDiff(new []string, old []string) ([]string, []string, []string) {
+	added := []string{}
+	removed := []string{}
+	unchanged := []string{}
+	for _, newVal := range new {
+		if !Contains(old, newVal) {
+			added = append(added, newVal)
+		} else {
+			unchanged = append(unchanged, newVal)
+		}
+	}
+	for _, oldVal := range old {
+		if !Contains(new, oldVal) {
+			removed = append(removed, oldVal)
+		}
+	}
+	return added, removed, unchanged
 }
 
 //StringOrNil returns a pointer to the input string, but returns nil if input is empty