+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Admin update account authorizations API [#484](https://github.com/rokwire/core-building-block/issues/484)
+- Set up admin create account endpoint [#365](https://github.com/rokwire/core-building-block/issues/365)
 - Prepare the project to become open source [#129](https://github.com/rokwire/core-building-block/issues/129)
 - Retrieve all service account tokens at once [#459](https://github.com/rokwire/core-building-block/issues/459)
 
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prevent admins from using service account management endpoints [#500](https://github.com/rokwire/core-building-block/issues/500)
 
 ### Changed
+- Get all admin level accounts [#486](https://github.com/rokwire/core-building-block/issues/486)
 - Update SECURITY.md [#504](https://github.com/rokwire/core-building-block/issues/504)
 
 ## [1.23.0] - 2022-04-26